Add warning about anonymous volume
The `Dockerfile` for the image uses an anonymous volume to persist the transformed config. This presents a problem in situations where a compose file is used. Volumes are not destroyed when changes are made to the compose file. This is in turn results in stale configuration. Additionally, it is rather difficult to debug when the configuration doesn't update as expected.